Modify fm-common deb pkg fles because there was missing fm_core
fm-common is a project that has both python and C code. The
first implementation of these rules and control files used the
debhelper automation to decide how to build and install the project.
Because of that only the C code was compiled and installed in the
deb package.

controller_config project depends on fm-common, and during it's
execution on a VM with Ubuntu 18.04 it was discovered that fm-common
package was missing fm_core subproject.

These committed changes solve the building and installation of
fm-common project into one deb fm-common-dev.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>